+++ noodles-cram/src/io.rs
@@ -2,9 +2,12 @@
 
 mod bit_reader;
 mod bit_writer;
+mod buf_reader;
 pub mod indexed_reader;
 pub mod reader;
 pub mod writer;
 
 pub(crate) use self::{bit_reader::BitReader, bit_writer::BitWriter};
-pub use self::{indexed_reader::IndexedReader, reader::Reader, writer::Writer};
+pub use self::{
+    buf_reader::BufReader, indexed_reader::IndexedReader, reader::Reader, writer::Writer,
+};